Expose entity details in MDSAL
Our previous implementation was very open about it state, which was
completely visible through RESTCONF (and from datastore).

We have a number of tools which need information relating to where each
entity lives -- and to that effect we want to expose a read-only API.

This patch instroduces odl-akka-eos model, which exposes a few RPCs,
which in turn provide information formerly visible from the datastore.

JIRA: CONTROLLER-1992
Change-Id: I85da8941f84ca9e109c4dec61fb07a697e937639
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>