Fix eos entity lookups with YangInstanceIdentifier
Caused by entity-name only being a string type
which was inadvertently trying to lookup all
entities as General.

Fix this by changing entity-name to a union of
instance-identifier and string.

Also changed up the output of get-entities slightly.
If the entity is string based it will only contain
the name in the output.
However if its Instance-identifier based it will contain
the entire identifier in the name field.

JIRA: CONTROLLER-2009
Change-Id: I93e108bd54b14850b6bf276ec1eca4d7671d7d87
Signed-off-by: Tomas Cere <tomas.cere@pantheon.tech>