Add timeout to limit repeated replication of Entries
Prevent the Leader from repeatedly sending AppendEntries
containing the same selection of Entries. This can lead to
out-of-memory issues on slower Followers. By the time the
Follower replies to the first AppendEntries, the Leader fills his
MessageQueue with a huge amount of AppendEntries messages, which
hold the same selection of Entries.

This patch adds a timeout which will be applied if the Leader tries
to send AppendEntries message with the same Entries as the last one.
Every time the timeout elapses and the Leader still tries to send the
same entries, the timeout gets multiplied by
repeated-replication-timeout-multiplier up to the maximum duration
specified by repeated-replication-max-timeout-seconds.
The timeout is reset whenever the Leader tries to send a different
selection of entries.

Signed-off-by: Tibor Král <tibor.kral@pantheon.tech>
Change-Id: Ic7753e4f8b1605ad33a08df89bedceb54a7915e4