Expose RaftState retrieval timeout config knob
The timeout is hard-coded to 10s. However in cases, when
the Shard Actor gets overloaded the request might fail
with AskTimeoutException. In this case the caller of
MBeanServer.getAttribute(shard, "RaftState") receives
NULL. In cases, when the slow-down is caused by limited
HW resources, having the option to tweak the value is
valuable.

Signed-off-by: Tibor Král <tibor.kral@pantheon.tech>
Change-Id: I1cd85567ebcdc21c5b2be07544c4169c9b1990af