Diagnose mockito downgrade better
Mockito is doing a very weird thing to get the inline MockMaker work,
which is known to not work with OSGi. Clean up the way we are
approaching byte-buddy and provide a pointer to the issue.

Change-Id: I87420710c5b6859e984fb72204826b91c7e4863a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 17a018384a969e748b79c1146b982f3b924c178d)