Remove ask-based protocol client
Remove the use-tell-based-protocol knob, effectively making it always
true.

This means DistributedDataStore cannot be instantiated, which leads to
its removal. That in turn makes all of DatastoreContext,
OperationLimiter and similar classes superfluous, so we remove those as
well.

A few classes are used to drive the shard backend in integration tests,
and hence those are moved to test sources.

JIRA: CONTROLLER-2054
Change-Id: Ie20b1c898576d3c89b70b34121310e58faddbf8e
Signed-off-by: bentom-binoy <bentom.binoy@infosys.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>