Unconditionally remove knownFrontends
As noted in the comment, known frontends end up being restored from
metadata when we unpause (or become leader for that matter).

The DatastoreContext property being used in this conditional is a
frontend thing anyway, so the only thing it could really affect is
tests.

The only case where this would be needed is when we are the leader, have
started persisting the DisableTrackingPayload, but it has not propagated
during pause/unpause.

That is okay, as the leader handover ensures that all locally-appended
entries are appended, so that the new leader just continues on.

Change-Id: Icf775823aed91c15f7a8f74fa2f07faee3a084a0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 3d0b84870481f03e4187bfbb7da89ff23e37516b)