WIP: Introduce AppendEntries.Entry
We need to reign in Payload lifecycle, for which us passing around
ReplicatedLogEntry is counter-productive, as it conflates state.

Since AppendEntries is always sent to remote, only the state passed
through serialization is relevant -- and we now hold that state in a
dedicated record.

This forces explicit transform when the two concepts meet.

// FIXME: we have an identity comparison here!

JIRA: CONTROLLER-2044
Change-Id: Iade4704d749ac595cf64a19cf122518e706f01b5
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>