Make Netty-3 dependency optional
Akka is using an long-obsolete version of Netty in for classic remoting.
While this is not a real problem by default, as the user needs to
configure it for the classes to be loaded, it is tripping up static
analysis tools.

This patch makes the org.jboss.netty.* import optional and removes our
packaging of Netty-3. This means that Akka Classic Remoting cannot be
configured unless the user also packages Netty-3 themselves.

JIRA: CONTROLLER-2119
Change-Id: I0a1bf611f140f2d5b6b83700893d0d69a3ba6115
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>