WIP: Use ScopedValue<Void> for completion offload
We are using ThreadLocal in its purest form here: we just need to know
if it is set.

ScopedValue gives use explicit control over that and exposes a
convenient isBound() boolean.

This allows us to better express what we need using a modern primitive
-- ditching use of Empty, as we no longer rely on the stored value. We
use Void, i.e. there is no point in accessing the value, as it can only
ever be null.

// FIXME: JIRA:

Change-Id: If392f47ce9c4bcf0ce5ad6f7edd5e47b0a181bc1
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>