WIP: Store cds-access-client generation in a file
We are storing the equivalent of three strings in Akka persistence,
which makes the state hard to reason about and hard to manipulate for
support purposes.

This patch provides an alternative by switching primary storage to
a property file in state/ directory.

Migration is automatic, in that if we encounter a persisted
ClientIdentifier, we generate the property file and also remember that
fact via a newly-introduced PersistenceTombstone.

Once we have a tombstone, we rely solely on the property file, which is
manipulated using atomic operations, to maintain this critical state.

// FIXME: finish up test suite

JIRA: CONTROLLER-2116
Change-Id: I05732d5640959c9a03c42b36e7c679782c91562a
Signed-off-by: Ruslan Kashapov <ruslan.kashapov@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>