Isolate Actor.getSender() references
We *really* want to use replyTo-in-request pattern, so that we can
disconnect actors from their behaviours (a la Akka Typed).

In order to do that, we must get a handle on which places are relying on
Actor.sender() and hence this patch.

We mass-migrate callers of Actor.sender() to Actor.getSender() and
override AbstractActor.getSender() to be deprecated for removal -- which
flags all callsites with a deprecation warning.

Change-Id: Ib62d3ce90e8a345ea54c88f226ba0507676ef89c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>