Introduce cluster.raft.spi.TermInfo
ElectionTerm is quite overloaded: it conflates the definition of
term+votedFor and lifecycle for updating it.

This patch introduces TermInfo as the holder of term+votedFor, leaving
ElectionTerm to be the lifecycle management part.

This ends up cleaning up quite a few interactions, as we usually
interchange atomic TermInfo, not its constituent parts.

JIRA: CONTROLLER-2127
Change-Id: I456a85dc96a01627c89a74d57520a8179dc9b4f9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>