Refactor sendSnapshotChunk()
Use Optional.isEmpty() to guard an early return. Then minimize the size
of the try/catch block, resuling in another possible early return.

Take advantage of the fields we are accessing, so that we get their
reuse -- which also improves clarity, as we use names that correspond to
InstallSnapshot fields.

JIRA: CONTROLLER-2074
Change-Id: Ibb2f6d9e6be726c2346f6b8358899eae56f7ac21
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>