Refactor SnapshotManager
The state keeping here is atrocious: the class is implemented as a
forwarder to current behavior and the actual implementations are hidden
behind a screen of default implementations.

This patch introduces SnapshotManager.Task as the core state-keeping
anchor. We then have a couple of records which hold state related to the
task.

This flushes out the fact we really have two different persisting paths,
depeding on what triggered it.

JIRA: CONTROLLER-2135
Change-Id: I9b667994d6f3d83ed3cffe3c7375978fb26d68c5
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>