Add cluster.raft.LocalAccess
Current layout of state management needs some better encapsulation, i.e.
something RaftActor sees from local storage perspective. One aspect
is TermInfoStore, as tied logId.

Introduce LocalAccess to maintain a combination of a TermInfoStore and a
logId.

Tests are migrated so that the test suites not involved with persistence
use the newly-improved TestTermInfoStore, or FailingTermInfoStore if
they want to assert that persistence was not called.

Common path with RaftActor is exposed via the
DataPersistenceProvider-based constructor, allowing those involving
persistence call out into that path to improve consistency.

JIRA: CONTROLLER-2133
Change-Id: I5960036b599f76b284d53d60bbdf8a58c706f290
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>