Reimplement TermInfoStore
This patch introduces storage of TermInfo in a property file rather than
in Akka/Pekko persistence. We also refactor the
RaftActor/RaftActorRecoverySupport surface to give recovery direct
access to AbstractPersistentActor, without indirection through
DataPersistenceProvider.

This patch is quite invasive as we need to push down the location of
state directory.

JIRA: CONTROLLER-2133
Change-Id: I24d296fc232fb62de30a8f00bdd5f381dc862115
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>