Clean up AbstractReplicatedLog
The string we get is tied to RaftActorContext.getId(), let's call it
logId, just as we do in LocalAccess.

Force subclasses to provide initial data, so that it is clear the
magic values we are using are for test/mock classes -- which now specify
them explicitly.

Use List.of() instead of Collections.emptyList() as a final touch.

JIRA: CONTROLLER-2073
Change-Id: I784d02ecc55b668e2d991db6d6ff4e109e033d52
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>