WIP: Eliminate StorageException
StorageException is a RuntimeException which is hiding potential failure
modes. Most of its use is wrapping an IOException, for which an
UncheckedIOException is a better match.

Eliminate StorageException and use IOException wherever a
StorageException would occur. sal-akka-segment-journal maps these
failures to UncheckedIOExceptions, so that these cause a terminal
failure of the actor.

StorageException.TooLarge becomes EntryTooBigException, derived from
IOException (and requiring EOFException as evidence).

StorageException.OutOfDiskSpace becomes OutOfDiskSpaceException, derived
from IOException as well.

// FIXME: four more call sites to reconcile

Change-Id: I2b4dece1db6cfc77fb988fa4a19b2ea7c25588c6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>