Clean up RaftActor name handling
We have three concepts here:
- persistenceId, which is left to be defined by subclasses
- id, which gets passed to the constructor
- Shard.name, which is effectively the same thing

Eliminate the need to call overridable methods in constructor by
codifying the constructor id as persistenceId(), which is now a final
method in RaftActor.

Also prevent RaftActor.getId() from being overridden: this is the id
passed to constructor, even when we ditch persistence. We also eliminate
Shard.name, as it was always the same thing as getId().

Change-Id: I007964f8c17c5693bd96a030c4b75821d491254b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>