Add ClusterConfig
RAFT paper mentions cluster configuration, which is what we have in
ServerConfigurationPayload.

Note that while we currently only transport ServerInfo, the structure is
flexible at the end of the day.

For now we call the ServerInfos we have serverInfo(), but really that
should be fused with the logic that uses it -- because this feels like
the List should be a Map and encapsulated in something called
ElectionPolicy or something.

We also remove PersistentPayload, as we are the only user -- and any
further access to being this special means you are part of
ClusterConfig, so explicit updates are in order.

JIRA: CONTROLLER-2137
Change-Id: Ie3eee97bbf71401889f835f2b8be9757c3f76c22
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>