Do not pass around logger and other clean ups
This patch performs a sweeep on sal-akka-raft logging, using per-class
loggers rather than passing one logger around.

AbstractLeader.printFollowerStates() is improved slightly to use fluent
StringBuilder append() calls.

Candidate.handleRequestVoteReply() is improved to be more linear in an
effort to make it clearer.

SnapshotTrackerTest is converted to JUnit5 to for accurate tracking of
which method call throws the expected exception.

Change-Id: I5e7142f2b1776da8edd21d302da55b1b0ab29d18
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>