Check listener state before forwarding notification
We have an intermittently-failing test because we are not checking
whether the registration is closed -- relying instead on the much slower
process of removing the listener.

Add the missing check, improving the turnaround on stopping event
delivery.

Change-Id: Ia51e5798dce8c322ca9d9a477ffae041ecffd5ef
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit efab8443338c829b1f04dd784296d2333ff44a9b)