WIP: Make snapshot create synchronous
The snapshot acquisition/serialization is lumped into a single
asynchronous call here, adding complexity and potential code
duplication.

We really should be acquiring an immutable state snapshot and have
RaftActor deal with its offloading internally, so that the interface
is simple and without special cases.

This ends up eliminating the need for GetSnapshotReplyActor, as it now
is a completely unneeded indirection.

JIRA: CONTROLLER-2134
Change-Id: Ic9e2fae68948be74ca1a861b4d45547bcf62a6ab
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>