Improve AbstractLeader class hierarchy
AbstractLeader.handleAppendEntriesReply() always results in 'this' being
returned and is overridden in the three subclasses.

Rename it to processAppendEntriesReply(), without the ability to change
behavior and make it final. The subclasses then use it as a common
utility, doing their own thing as needed.

Since Leader is subclassed in tests, we lock down all its methods except
the single one that is being overridden.

Change-Id: I594bebedfa612f7e946d97040ea55c053c2c9f3d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>