Do not reset ReplicatedLog
Replacing ReplicatedLog is a rather bad thing, as we need to be mindful
when we can and cannot cache it.

This patch takes the first step towards making it an invariant:
introduce a resetToSnapshot() method which resets the state and take
advantage of that.

We also deprecate setReplicatedLog() for removal.

JIRA: CONTROLLER-2137
Change-Id: Ifc9ce0b0910214d5d0f314f26623fc4583fd1849
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>