Promote RaftState
Let's start a new package, opendaylight.raft.api, which holds a single
enumeration for now.

Since this is a plain enumeration, let's name it 'RaftRoles', acting as
a repository, which is how we are using it: RaftRoles.Leader, etc.

This frees up 'RaftState' for something actually holding state.

JIRA: CONTROLLER-2134
Change-Id: I50b25b4d13cf426285face63ed8be2a8356ab83d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>