WIP: Add RaftState
The lifecycle of RaftActor is a bit weird, with various bits and pieces
wired in quite contrived ways.

We essentially want the state of a RaftActor to be properly contained,
which we sort of do in RaftActorContext, but that is incomplete and the
context is a god object as is.

This patch introduces RaftState to hold the state and performs explicit
transitions

// FIXME: need to complete this through to having an StartedState, which
//        is where we ditch all stuff related to recovery and operate
//        with RaftActorBehaviors

JIRA: CONTROLLER-2137
Change-Id: If60022f9ad7cf357b48f8bd2a77af3a5adfa07eb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>