Add cluster.raft.spi.SnapshotFileFormat
This is a first cut at having a sensible file format for storing
snapshots. Unlike Pekko-based persistence, which stored the complete
contents of a Snapshot, we just pick store data that cannot be inferred
from the way things are laid out.

The following o.o.raft.spi constructs are introduced:
- InstallableSnapshot, a handle on byte state that can be tranferred via
  InstallSnapshot RPCs
- SizedStreamSource, the moral equivalent of Guava's ByteSource
- FileStreamSource, a SizedStreamSource served from a section of a file

JIRA: CONTROLLER-2134
Change-Id: I9743ec8b701badd12857e483acefb95709d9b68a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>